requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
Flask_Cors==4.0.0
mysql_connector_repackaged==0.3.1
Requests==2.31.0
schedule==1.2.1
selenium==4.17.2
